@@ -1,12 +1,11 @@
 #ifndef NGLSCENE_H_
 #define NGLSCENE_H_
-#include <QOpenGLWindow>
 #include <ngl/Mat4.h>
 #include <ngl/Transformation.h>
 #include <ngl/Text.h>
 #include "WindowParams.h"
 #include <memory>
-
+#include <QOpenGLWindow>
 //----------------------------------------------------------------------------------------------------------------------
 /// @file NGLScene.h
 /// @brief this class inherits from the Qt OpenGLWindow and allows us to use NGL to draw OpenGL
